+++ oidc-controller/api/core/models.py
@@ -1,7 +1,23 @@
-import uuid as uuid_pkg
 from datetime import datetime
 
+from bson import ObjectId
 from pydantic import BaseModel, Field
+
+
+class PyObjectId(ObjectId):
+    @classmethod
+    def __get_validators__(cls):
+        yield cls.validate
+
+    @classmethod
+    def validate(cls, v):
+        if not ObjectId.is_valid(v):
+            raise ValueError("Invalid objectid")
+        return ObjectId(v)
+
+    @classmethod
+    def __modify_schema__(cls, field_schema):
+        field_schema.update(type="string")
 
 
 class HealthCheck(BaseModel):
@@ -16,7 +32,10 @@ class StatusMessage(BaseModel):
 
 
 class UUIDModel(BaseModel):
-    uuid: uuid_pkg.UUID = Field(default_factory=uuid_pkg.uuid4)
+    id: PyObjectId = Field(default_factory=PyObjectId, alias="_id")
+
+    class Config:
+        json_encoders = {ObjectId: str}
 
 
 class TimestampModel(BaseModel):
