+++ oidc-controller/api/verificationConfigs/router.py
@@ -4,7 +4,6 @@ from ..core.models import StatusMessage
 
 from .crud import VerificationConfigCRUD
 from .models import (
-    VerificationConfigCreate,
     VerificationConfigPatch,
     VerificationConfigRead,
     VerificationConfig,
@@ -17,6 +16,7 @@ router = APIRouter()
     "/",
     response_description="Add new verification configuration",
     response_model=VerificationConfig,
+    response_model_exclude_unset=True,
 )
 async def create_ver_config(ver_config: VerificationConfig):
     return await VerificationConfigCRUD.create(ver_config)
@@ -24,8 +24,9 @@ async def create_ver_config(ver_config: VerificationConfig):
 
 @router.get(
     "/{ver_config_id}",
-    response_model=VerificationConfigRead,
     status_code=http_status.HTTP_200_OK,
+    response_model=VerificationConfigRead,
+    response_model_exclude_unset=True,
 )
 async def get_ver_conf(
     ver_config_id: str,
@@ -35,8 +36,9 @@ async def get_ver_conf(
 
 @router.patch(
     "/{ver_config_id}",
-    response_model=VerificationConfigRead,
     status_code=http_status.HTTP_200_OK,
+    response_model=VerificationConfigRead,
+    response_model_exclude_unset=True,
 )
 async def patch_ver_conf(
     ver_config_id: str,
@@ -47,8 +49,8 @@ async def patch_ver_conf(
 
 @router.delete(
     "/{ver_config_id}",
-    response_model=StatusMessage,
     status_code=http_status.HTTP_200_OK,
+    response_model=StatusMessage,
 )
 async def delete_ver_conf_by_uuid(
     ver_config_id: str,
